refactor(task): 收敛 C1 分层修复为 claim_join_gate 本地最小 Protocol
task 模块消费点专属:仅 claim_join_gate 用到 get_config/set_config 两法,无需把共享
SystemConfigServiceProtocol 下沉到 plugin_api(独立 system_config.py)+ api re-export。

- claim_join_gate 内定义 @runtime_checkable _SystemConfigStore(仅 get_config/set_config),
  __init__ 标注为 _SystemConfigStore | None,@inject 不变。
- 撤回 plugin_api/system_config.py + api/system_config_service.py re-export +
  api/README 中 plugin_api.system_config 声明;claim_join_gate 不再引 api/plugin_api,
  core→api 分层门禁合规。
- provider(task_module.task_claim_join_gate)显式 config=injector.get(SystemConfigServiceProtocol)
  构造注入,无 typed-resolve;SystemConfigService 结构满足本地 Protocol,DI 无需重绑。
- 顺手修正 E402(import 紧凑置顶,Protocol 类下移)。

验证:arch(core→api)/module boundaries + task runner integration 回归 + claim-join-filter
endpoint 均绿;ruff clean;全 src 无 plugin_api.system_config 残留引用。
